Rename the directory controller into reconciler
This PR is the the first step of the refactoring to use knative/pkg.
