+++ pkg/reconciler/knativeserving/add_minikube.go
@@ -13,21 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package controller
+package knativeserving
 
 import (
-	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"github.com/knative/serving-operator/pkg/reconciler/knativeserving/minikube"
 )
 
-// AddToManagerFuncs is a list of functions to add all Controllers to the Manager
-var AddToManagerFuncs []func(manager.Manager) error
-
-// AddToManager adds all Controllers to the Manager
-func AddToManager(m manager.Manager) error {
-	for _, f := range AddToManagerFuncs {
-		if err := f(m); err != nil {
-			return err
-		}
-	}
-	return nil
+func init() {
+	platforms = append(platforms, minikube.Configure)
 }
