+++ pkg/reconciler/knativeserving/common/extensions.go
