+++ pkg/reconciler/knativeserving/common/images.go
