+++ pkg/reconciler/knativeserving/common/images_test.go
